PHP 7.2 compatibility for DataSet / DataGrid
Prior to PHP 7.2, `count()` worked on scalar values, always returning `0` for `null`, and `1` for strings and numbers.
In PHP 7.2, it is no longer the case; `count()` expects an array or a `Countable` implementation:

```sh
$ php -r 'count(null);'
PHP Warning:  count(): Parameter must be an array or an object that implements Countable in Command line code on line 1
```

The package used `count()` on `DataSet::$orderby` in `DataSet::onOrderby()`, called from the `rapyd::datagrid` view.
Since the default value for `DataSet::$orderby` is `null`, the grid no longer worked in PHP 7.2, unless the default order is set explicitly (i.e. with `$grid->orderBy('field', 'direction')`).

This commit changes that to `!empty()` to make it compatible with PHP 7.2.